+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - Fixed: Fix build on musl by disabling, for now, reflink copies.
 
+- Fixed: Don't error on diffs that contain only git file moves, by moving to Flickzeug. Thanks to @eirnym.
+
 ## 26.0.0 2025-12-07
 
 - Changed: The default is now *not* to shuffle mutants: they run in the deterministic order they are generated in the source tree. This should give somewhat better locality of reference due to consecutively testing changes in each package or module. The previous behavior can be restored with `--shuffle`.